requirements.txt
anyio==4.4.0
black==24.8.0
certifi==2024.7.4
click==8.1.7
environs==11.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
marshmallow==3.22.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
python-telegram-bot==21.4
sniffio==1.3.1
